requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.7.0
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
coqpit-config==0.1.2
coqui-tts==0.25.1
coqui-tts-trainer==0.2.0
cutlet==0.4.0
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
diskcache==5.6.3
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
fugashi==1.4.0
future==1.0.0
grpcio==1.68.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
HyperPyYAML==1.2.2
idna==3.10
inflect==7.5.0
jaconv==0.4.0
Jinja2==3.1.5
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
llama_cpp_python==0.3.5
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mecab-python3==1.0.10
mojimoji==0.0.13
monotonic-alignment-search==0.1.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.11
networkx==3.4.2
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.1
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.7
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.2
srsly==2.5.0
SudachiDict-core==20241021
SudachiPy==0.6.9
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers==4.46.2
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.2
unidic-lite==1.0.8
urllib3==2.3.0
voicevox-client==0.4.1
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
whisper-openai==1.0.0
wrapt==1.17.0
yarl==1.18.3
